feat(core): add useLanguageHelper hooks
